screenshot-v1
A screenshot tool written by Python.You can use it to make screenshots and hang it at the top of your desktop.You can also store the location of all the screenshots,as well as return to the status just as when you stored them. You can also flip or rotate the images. The storage path and the shortcut keys can be set in config.xml.